now only asks are draggable. but ok ugh maybe i need something more full-featured after all. gonna try jqueryui.
my problem is this: i can't tell what i'm dragging to, because i don't
know how to get the mouseover event to propagate past the element being
dragged down to the element below. and if i disable pointer-events on
the dragged object, it can't tell when it's released, so it "sticks" to
the cursor. the possible solutions generally involve a lot of cycling
and… and i don't really get it. soooo i guess i'd rather have a black
box that works than a gray box that only sorta does. :-/